Move InfoBox to components/ui and add storybook
