+++ airbyte-webapp/src/components/ui/InfoBox/index.ts
